+++ calypso_anemometer/cli.py
@@ -1,4 +1,3 @@
-import json
 import logging
 import os
 import sys
@@ -6,7 +5,7 @@ import sys
 import click
 
 from calypso_anemometer.core import CalypsoDeviceApi
-from calypso_anemometer.util import make_sync, setup_logging
+from calypso_anemometer.util import make_sync, setup_logging, to_json
 
 logger = logging.getLogger(__name__)
 
@@ -35,8 +34,10 @@ async def info(ctx):
     try:
         async with CalypsoDeviceApi(ble_address=os.getenv("CALYPSO_ADDRESS")) as calypso:
             device_info = await calypso.get_info()
-            print(json.dumps(device_info.asdict(), indent=2))
     except Exception as ex:
+            print(to_json(device_info))
+            device_status = await calypso.get_status()
+            print(to_json(device_status.aslabeldict()))
         logger.error(ex)
         sys.exit(1)