+++ CHANGES.rst
@@ -11,6 +11,7 @@ in progress
 - Implement client interface as context manager
 - Increase default timeout values to 15 seconds
 - Rework device info acquisition
+- Read and decode device status bytes: mode, rate, compass
 
 
 2022-xx-xx 0.0.0